optimize_videos: Add flags to exit on error and unset variables

+++ programs/optimize_videos.sh
@@ -21,6 +21,9 @@
 #
 # ⚠ Warning: this tool overwrites the source videos! Backup your videos before using this program.
 
+# Exit if there's any unset variable
+set -eu
+
 NAME="optimize_videos.sh"
 ERROR_RESIZE="Incorrect value for --resize. Turning resize off."
 TUNE_OPTIONS="film, animation, grain, stillimage"
@@ -116,7 +119,7 @@ compress_videos() {
 		path_temp=$(dirname "$filepath")/temp_$(echo "$filename" | sed 's/\.[A-Za-z0-9]+$/.mp4/')
 		path_out=$(echo "$path_temp" | sed 's/temp_//')
 		ffmpeg_command=$(printf "ffmpeg $args \"%s\"" "$filepath" "$path_temp")
-		test $is_dry_run -eq 0 && eval "$ffmpeg_command" || echo "$ffmpeg_command"
+		test $is_dry_run -eq 0 && eval "$ffmpeg_command" || echo "$ffmpeg_command" && continue
 		test $is_dry_run -eq 0 && mv -v "$path_temp" "$path_out" || echo Moving "$path_temp" to "$path_out"
 	done <"$1"
 }
